(tokenizer) Add Gemma-3-270m w/ uint32 tokens
Switch from LLaMA to Google Gemma tokenizer (262k vocab) requiring
uint32 token data type instead of uint16. Add HF_TOKEN support for gated
model authentication.

- Update tokenizer from LLaMA to google/gemma-3-270m in hparams.json
- Change SharedShardedDataset and ShardedDatasetManager to uint32
- Add explicit token_dtype=np.uint32 in miner.py and validator.py
- Add HF_TOKEN environment variable support in hparams.py
- Keep test config with smaller huggyllama/llama-7b tokenizer
- Update miner.md and validator.md documentation:
  - Add HF_TOKEN as required environment variable
  - Fix dataset reference typo (shared_sharded_datset.md)
  - Update dataset references to shared_sharded_dataset.md

+++ src/tplr/sharded_dataset.py
@@ -45,7 +45,7 @@ class SharedShardedDataset(Dataset):
         rank: int,
         world_size: int,
         *,
-        token_dtype: npt.DTypeLike = np.uint16,  # MUST match preprocessing
+        token_dtype: npt.DTypeLike = np.uint32,  # MUST match preprocessing
         file_prefix: str = "train",  # Allow custom file prefix (e.g., "val", "eval")
     ):
         super().__init__()
@@ -167,7 +167,7 @@ class ShardedDatasetManager:
         rank: int,
         world_size: int,
         comms: tplr.comms.Comms,
-        token_dtype: npt.DTypeLike = np.uint16,
+        token_dtype: npt.DTypeLike = np.uint32,
         file_prefix: str = "train",
     ):
         """Initializes the dataset manager.
